update unsigned id -> long id

+++ code/include/creatures.h
@@ -9,7 +9,7 @@
     Bestiaire *initModalBestiary(ListeCompetence *modalCreaturesSkills);
     int generateCreatureInBestiary(Bestiaire *modalBestiary, Bestiaire *bestiary);
     Bestiaire *initRandomBestiaryFromDangerosityGroupLevel(Bestiaire *modalBestiary, int dangerosityLevel);
-    int addCreatureInBestiary(Bestiaire *modalBestiary, Bestiaire *bestiary, unsigned idConf);
+    int addCreatureInBestiary(Bestiaire *modalBestiary, Bestiaire *bestiary, long idConf);
     int setDeathStateCreature(CreatureMarine *creature);
     void freeBestiary(Bestiaire *bestiary);
     void freeBestiaryContent(Bestiaire *bestiary);
